# astroneko-backend/docs/swagger.yaml
basePath: /
definitions:
  agent.ClearStateRequest:
    properties:
      session_id:
        type: string
    type: object
  agent.ClearStateResponse:
    properties:
      status:
        type: string
    type: object
  agent.ReplyRequest:
    properties:
      session_id:
        type: string
      text:
        type: string
      user_id:
        type: string
    required:
    - text
    type: object
  agent.ReplyResponse:
    properties:
      card:
        type: string
      meaning:
        type: string
      message:
        type: string
      session_id:
        type: string
      status:
        type: string
    type: object
  astro_boxing_waiting_list.AstroBoxingWaitingListUserResponse:
    properties:
      created_at:
        type: string
      email:
        type: string
      id:
        type: string
      updated_at:
        type: string
    type: object
  astro_boxing_waiting_list.IsInAstroBoxingWaitingListResponse:
    properties:
      is_in_waiting_list:
        type: boolean
    type: object
  astro_boxing_waiting_list.JoinAstroBoxingWaitingListRequest:
    properties:
      email:
        type: string
    required:
    - email
    type: object
  astro_boxing_waiting_list.JoinAstroBoxingWaitingListResponse:
    properties:
      currentPage:
        type: integer
      data: {}
      latestAiSuggestQuestions:
        items:
          type: string
        type: array
      meta:
        additionalProperties: true
        type: object
      perPage:
        type: integer
      status:
        $ref: '#/definitions/shared.Status'
      totalItem:
        type: integer
    type: object
...
  shared.ResponseBody:
    properties:
      currentPage:
        type: integer
      data: {}
      latestAiSuggestQuestions:
        items:
          type: string
        type: array
      meta:
        additionalProperties: true
        type: object
      perPage:
        type: integer
      status:
        $ref: '#/definitions/shared.Status'
      totalItem:
        type: integer
    type: object
  shared.Status:
    properties:
      code:
        type: string
      errorID:
        type: string
      message:
        items:
          type: string
        type: array
      successID:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: REST API for Astroneko Backend
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Astroneko Backend API
  version: "1.0"
paths:
...
  /v1/api/agent/clear-state:
    post:
      consumes:
      - application/json
      description: Clear the conversation state for the cat fortune agent
      parameters:
      - description: Clear state request (user_id will be set from auth)
        in: body
        name: clear_state
        schema:
          $ref: '#/definitions/agent.ClearStateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agent.ClearStateResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      security:
      - BearerAuth: []
      summary: Clear agent state for authenticated user
      tags:
      - agent
  /v1/api/agent/reply:
    post:
      consumes:
      - application/json
      description: Send a message to the cat fortune agent and receive a response.
        Works for both authenticated users (unlimited) and guests (3 requests/day).
        For authenticated users, user_id is automatically extracted from auth token.
        For guests, session fingerprint is used. session_id is optional.
      parameters:
      - description: Message to send to agent (session_id is optional)
        in: body
        name: reply
        required: true
        schema:
          $ref: '#/definitions/agent.ReplyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/agent.ReplyResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "429":
          description: Rate limit exceeded for guest users
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      security:
      - BearerAuth: []
      summary: Send message to agent and get reply
      tags:
      - agent
  /v1/api/astro-boxing-waiting-list/check:
    get:
      consumes:
      - application/json
      description: Check if an email exists in the astro boxing waiting list
      parameters:
      - description: Email to check
        in: query
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/astro_boxing_waiting_list.IsInAstroBoxingWaitingListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      summary: Check if email is in astro boxing waiting list
      tags:
      - astro_boxing_waiting_list
  /v1/api/astro-boxing-waiting-list/join:
    post:
      consumes:
      - application/json
      description: Add user to the astro boxing waiting list
      parameters:
      - description: Join request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/astro_boxing_waiting_list.JoinAstroBoxingWaitingListRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/astro_boxing_waiting_list.JoinAstroBoxingWaitingListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      summary: Join astro boxing waiting list
      tags:
      - astro_boxing_waiting_list
  /v1/api/astro-boxing-waiting-list/users:
    get:
      consumes:
      - application/json
      description: Get paginated list of astro boxing waiting list users
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/astro_boxing_waiting_list.AstroBoxingWaitingListUserResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      summary: Get astro boxing waiting list users
      tags:
      - astro_boxing_waiting_list
  /v1/api/astro-boxing-waiting-list/users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user from astro boxing waiting list by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/shared.ResponseBody'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/shared.ResponseBody'
      summary: Delete astro boxing waiting list user
      tags:
      - astro_boxing_waiting_list
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"